pyproject.toml
[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B"]
exclude = [".venv", "__pycache__", "build", "dist"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project]
name = "ai-draw-guess"
version = "0.1.0"
description = "AI Draw & Guess game"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "python-dotenv",
    "pydantic",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
